[other] fix ros bug in uint8_t

+++ libhesai/driver_param.h
@@ -92,7 +92,7 @@ typedef struct DecoderParam
   float frame_frequency = 0;
   float default_frame_frequency = DEFAULT_MAX_MULTI_FRAME_NUM;
   bool update_function_safety_flag = false;
-  uint8_t echo_mode_filter = 0;
+  uint16_t echo_mode_filter = 0;
 } DecoderParam;
 
 ///< The LiDAR input parameter
